requirements.txt
argparse==1.2.1
colored==1.2.1
kafka-python==1.0.2
six==1.10.0
wsgiref==0.1.2